delete key prefix token select, change files names, GrupedTokenTypes and export select components into different components
